+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changes
 
 * Improve support for IDE/editor plugins (.e.g vscode, sublime, intellij-idea)
+* Add `.gitignore` file when creating new projects
 
 ### Bugs fixed
 
@@ -393,6 +394,7 @@
 ### 改进
 
 * 改进对IDE和编辑器插件的集成支持，例如：Visual Studio Code, Sublime Text 以及 IntelliJ IDEA
+* 当生成新工程的时候，自动生成一个`.gitignore`文件，忽略一些xmake的临时文件和目录
 
 ### Bugs修复